fix: batch resolve 7 design-decision issues
closes #70  — vault delete now cleans up chunks/embeddings before removing notes
closes #64  — trash permanent delete cleans up embeddings and chunks
closes #118 — delete orphaned S3 object when attachment DB insert fails
closes #77  — share endpoint copies S3 content instead of referencing original key
closes #110 — rename rateLimit.js to loginLockout.js for clarity
closes #56  — migration runner now records applied migrations in schema_migrations table
closes #91  — hash PII (emails) in log output, remove names/phone from contact logs
